# ORC-1020: [C++] Optimize RleDecoderV2::nextDirect base on bit sizes