… c++/test/TestRleDecoder.cc